% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEBdata-package.R
\docType{package}
\name{DEEBdata-package}
\alias{DEEBdata}
\alias{DEEBdata-package}
\title{DEEBdata: Create a data for a simulation study to compare differential equation estimators}
\description{
This package allows to create a data base of simulated data for evaluating and comparing methods for estimation of differential equations.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
