% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEData-package.R
\docType{package}
\name{DEEData-package}
\alias{DEEData}
\alias{DEEData-package}
\title{DEEData: Create a data for a simulation study to compare differential equation estimators}
\description{
This package allows to create a data base of simulated data for evaluating and comparing methods for estimation of differential equations.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
